backend/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["msklnkk <128256503+msklnkk@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.2"
pydantic-settings = "^2.5.2"
uvicorn = "^0.31.1"
asyncpg = "^0.29.0"
greenlet = "^3.1.1"
alembic = "^1.13.3"
tenacity = "^9.0.0"
pyjwt = "^2.10.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.19"
pydantic = {extras = ["email"], version = "^2.12.4"}
bcrypt = "3.2.2"
passlib = {version = "1.7.4", extras = ["bcrypt"]}
requests = "^2.32.5"
aiohttp = "^3.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
